// types.ts
export interface AcademicEvent {
  id: string;
  date: string; // ISO string YYYY-MM-DD
  endDate?: string; // ISO string YYYY-MM-DD for multi-day events
  title: string;
  type: 'pembelajaran' | 'libur' | 'lainnya' | 'mgbk' | 'ujian' | 'libur_sekolah' | 'dinas_dalam' | 'dinas_luar' | 'pelatihan';
  description?: string;
  reminder?: boolean;
}